feat: delete unecessary classes + move back logic to TokenStoreKeyValue
